* need to fix test suit for -O0 vs -O2  ???
the test pass in -O0 mode but not in -O2
In fact the rounding is not the same between -O0 and -O2
so when I sort the geometries, there is a difference.

Now I also check thah the distance between two geometries is greater
than HKL_EPSILON to sort them.

that way I avoid this rounding problem present on the 13th digits.