Rework the unit part of the hkl library
Prepare the final API for the inclusion of the udunits library.
This will allow to set the user unit prefered output of
the instrumented methods.

until now this was mostly used for radian to degree convertion.
But we can imagine other convertions in the futur.

the binding API also changed