(wip) rework of the pseudo axis part
now the set method return the HklGeometryList.
It is now the responsability of the API user to release the memory
by calling hkl_geometry_list_free on the returned geometries.

fix the hkl-pseudoaxis-e6c petra2_3 test by using hkl_method_set_hkl_real
for the psi_constant_vertical_mode_operations.
This setter compute also a set of solutions by using the z-axis strategie
which consist in a rotation of the sample around the latest axis and a fit
of the detector position.

move the hkl_pseudo...randomize methodes into the hkl tap library instead of
the public API.

- Add the HklEngine Capabilities
  you can get them with hkl_engine_capabilities_get
  this tells you if an HklEngine is READABLE, WRITABLE, INITIALIZABLE

- add a new hkl_engine_initialized_get/set
  if an HklEngine is INITIALIZABLE you can set the value to TRUE.
  depending on the Mode, this can changes the mode parameters, save the
  diffractometer current position, ...