replace g_return_val_if_fail by hkl_error when relevant
In case of programming error the programm stop with an abort.
Using gdb you can obtain a backtrace with 'bt' command.
This hkl_error is like assert except that it is not possible to get rid
of them with NDEBUG.

Recoverable errors (not due to programming error) are reported via the
usual GError system.