add HKL_WARN_UNUSED_RESULT for the public API
if an hkl user does not use a return value from a method marked with this
attribute, a warning is displayed during the compilation.

for now only the method throwing a GError are instrumented. It is a good
reminder to deal with the GError.

BEWARE this warn is not activated for the libhkl compilation itself
using the CFLAGS -Wno-unused-result.