add an --enable-logging to activate the logging in the hkl library
For now this is optional to avoid the perfomance impact.
I do not have numbers for now but I prefer to make this optional in a first time.

I am not really happy with the implementation of the logging.
Only one method is instrumented for now

hkl_engine_pseudo_axes_values_set